Server: `MatchManager::DestroyMatch()`: Use `find_if` instead of `remove_if`
Since we want to access some properties of the match that is about to be removed, we should use `find_if` to get the iterator.

Using `remove_if` leaves the match object in an unspecified state, which, for example, caused a weird memory access issue when trying to destroy a match with 3 actual players.