Server: Let computer players interact with Spades and Hearts matches
Adds and organizes the code in such a way that allows computer players to interact with Spades and Hearts matches (the only ones where they're supported for now) when required.

Additionally, Spades matches now track whether Spades have been unlocked, which therefore allow tricks to be led with Spades. If not unlocked, that is now forbidden.

TODO: Actual bid/pass/card strategy logic for Spades and Hearts.