Server: Utilize non-blocking sockets for players
The `Socket` class now supports a non-blocking socket mode, which is used for both Win7 and WinXP player sockets. A send queue was implemented where raw data to be sent is saved until the send event is processed, or until the target socket becomes writeable. This means that `Socket::SendData()` will no longer block the thread.

Both `Win7::PlayerSocket` and `WinXP::PlayerSocket` are now responsible for implementing their own protocols, which means that there are no longer such helper functions in the `Socket` class - only basic `ReadData(char*, int)` and `SendData(char*, int)` are available for simplicity.

`Win7::PlayerSocket` now continuously fetches data over the socket until it receives data which contains "\r\n", ensuring that lines in the protocol will always be fully fetched and parsed individually. It's a way more mature way of navigating the protocol than relying on `recv()` giving us all the data at once.

`WinXP::PlayerSocket` no longer calls `recv()` to receive each and every message it requires. Instead, any message data is read through the network in `AwaitIncomingGenericMessage()`, keeping in mind the required length specified in the `MsgBaseGeneric` (base message header) itself, which is fetched separately before that. This means that the match no longer calls the `PlayerSocket` to fetch data from the network, thus hanging - it only asks the `PlayerSocket` to read the data from its message buffer, ensuring that this also doesn't block the match thread.

With these changes in mind, the match thread in `WinXP::Match::ProcessIncomingGameMessage()` is no longer prone to being stalled by a slow socket, which lead to the danger of the thread owning the mutex for too long and thus the mutex timing out and crashing the server.