Server: Allow hosting a simple HTTP server which serves an HTML about waiting lobbies
A new "enablehttp" option was added, which makes the server application create another listening socket on port 80 for a simple HTTP server. It only serves a root HTML page, which contains a table with the amount of waiting lobbies.

The waiting lobbies page separates lobbies, based on match skill level, if "skiplevel" is not set. Otherwise, a single counter is only shown on version/game. This page aims to make it easy for players to find others to play with.